requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.3.1
django-filter==24.1
django-heroku==0.3.1
djangorestframework==3.15.0
gunicorn==22.0.0
Markdown==3.6
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
Pygments==2.17.2
PyJWT==2.8.0
python-barcode==0.15.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
xlwt==1.3.0
